Docker install script, api env vars, IP script, dump script (#10)
* docker run script

* docker scripting, environment vars.

created a docker script to delete and rebuild api image and containers. also added env vars for dev and prod services seperately. and AWS env for the api key, hopefully working.

* create pool

* default user image + fromeroad db backups

* updated to use .sql files for mysql docker

moved sql files to another folder. created /dev/ /prod/ to store .sql dump file
made docker use this dump file for data on creation so now data will work when deleted and remade

* update data/default/ and removed deletion mysql

* detecting ip change and updating

needs to be automated for when net reboots

* Create update_route53_script.ps1

* updated to work

needs to be called by .bat file but that is stored on laptop so that the script cannot be edited for malicious reasons

* ignore prod data, hourly backups.