better jwt verify. delete comments, prep getUser
made it so tokens are validated against the userID that sent the request. This means people cant use their own token and someone elses ID in postman to change their post or smt.

changed delete comments logic so it is instant rather than making another api call.

prepped getuser so when people go on profiles the password isnt passed back.